ioutil.ReadDir calls stat on every directory entry. These calls are not required
to count the number of file descriptors.

For a process with many open file descriptors this can be costly.

In a local benchmark on a directory with 10k files, we observe the following improvement by using (*os.File).Readdirnames directly:

```
goos: linux
goarch: amd64
pkg: github.com/elastic/gosigar
cpu: Intel(R) Xeon(R) CPU @ 2.80GHz
                    │   old.txt    │               new.txt               │
                    │    sec/op    │   sec/op     vs base                │
GetDirEntryCount-24   23.874m ± 0%   2.967m ± 0%  -87.57% (p=0.000 n=25)

                    │    old.txt    │               new.txt                │
                    │     B/op      │     B/op      vs base                │
GetDirEntryCount-24   3310.4Ki ± 0%   729.4Ki ± 0%  -77.97% (p=0.000 n=25)

                    │   old.txt   │               new.txt               │
                    │  allocs/op  │  allocs/op   vs base                │
GetDirEntryCount-24   30.03k ± 0%   10.02k ± 0%  -66.62% (p=0.000 n=25)
```